refactor: 카드 오픈 전략패턴 패키지 경로 이동
